src/presentation/helpers/validators/email-validation.ts
import { InvalidParamError } from "../../errors"
import { EmailValidator } from "../../protocols/email-validator"
import { Validation } from "./validation"

class EmailValidation implements Validation {
  private readonly fieldName: string
  private readonly emailValidator: EmailValidator

  constructor(fieldName: string, emailValidator: EmailValidator) {
    this.fieldName = fieldName
    this.emailValidator = emailValidator
  }

  validate(input: any): any {
    const isValidEmail = this.emailValidator.isValid(input[this.fieldName])

    if (!isValidEmail) {
      return new InvalidParamError(this.fieldName)
    }
  }
}

export { EmailValidation }
